Supports loading file patterns from an external file
Patterns are now loaded from Patterns.json file inside the
user's AppData folder when the plugin is first invoked.
I also had to refactor the code a bit to simplify how the
JSON file would be structured.
Some additional refactoring was done so that the pattern
matching happens only on the file name, rather than the full
relative path to the file.