FIleName prompt improvements and minor bug fix
File Name prompt will now include a default filename with either .vb or .cs as extention depending on the project type. The filename part (except the extension) comes preselected so that the user can overwrite that part simply by typing. If the user types a '.' the extension gets selected so that the user can continue to overwrite the extension as well.
Also fixed the bug where non-templated files were creating duplciated nested folders.